backend/requirements.txt
# FastAPI and server
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.25
asyncpg==0.29.0
psycopg2-binary==2.9.9
alembic==1.13.1
pgvector==0.2.4

# LangChain
langchain==0.1.4
langchain-google-genai==0.0.6
langchain-community==0.0.16
langchain-text-splitters

# Document Processing
pypdf==4.0.1
python-docx==1.1.0

# Google Gemini
google-generativeai==0.3.2

# Vector DB - Qdrant
qdrant-client

# Telegram Bot
python-telegram-bot==20.7

# Utilities
python-dotenv==1.0.0
pydantic==2.5.3
pydantic-settings==2.1.0
aiofiles==23.2.1

# HTTP Client
httpx==0.26.0
